+++ app/controllers/application_controller.rb
@@ -10,6 +10,8 @@ class ApplicationController < ActionController::Base
   before_action :store_location
   before_action :set_cache_buster
 
+  layout "full_page"
+
   def store_location
     if !request.post? && !request.xhr? && params[:format] != "atom"
       if internal_action?(params[:controller], params[:action])